refactor instanceService to clusterMaintenanceService
