+++ helix-rest/src/test/java/org/apache/helix/rest/clusterMaintenanceService/TestMaintenanceManagementService.java
@@ -1,4 +1,4 @@
-package org.apache.helix.rest.server.service;
+package org.apache.helix.rest.clusterMaintenanceService;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -18,14 +18,12 @@ package org.apache.helix.rest.server.service;
  * specific language governing permissions and limitations
  * under the License.
  */
-
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 import org.apache.helix.AccessOption;
@@ -48,7 +46,6 @@ import org.mockito.MockitoAnnotations;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-
 import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyList;
@@ -60,7 +57,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
-public class TestInstanceService {
+
+public class TestMaintenanceManagementService {
   private static final String TEST_CLUSTER = "TestCluster";
   private static final String TEST_INSTANCE = "instance0.linkedin.com_1235";
 
@@ -70,7 +68,6 @@ public class TestInstanceService {
   private ConfigAccessor _configAccessor;
   @Mock
   private CustomRestClient _customRestClient;
-
   @BeforeMethod
   public void beforeMethod() {
     MockitoAnnotations.initMocks(this);
@@ -79,23 +76,35 @@ public class TestInstanceService {
     when(_configAccessor.getRESTConfig(TEST_CLUSTER)).thenReturn(restConfig);
   }
 
+  class MockMaintenanceManagementService extends MaintenanceManagementService {
+    public MockMaintenanceManagementService(ZKHelixDataAccessor dataAccessor,
+        ConfigAccessor configAccessor, CustomRestClient customRestClient, boolean skipZKRead,
+        boolean continueOnFailures, String namespace) {
+      super(dataAccessor, configAccessor, customRestClient, skipZKRead, continueOnFailures,
+          namespace);
+    }
+    @Override
+    protected Map<String, Boolean> getInstanceHealthStatus(String clusterId, String instanceName,
+        List<HealthCheck> healthChecks) {
+      return Collections.emptyMap();
+    }
+  }
+
   @Test
   public void testGetInstanceStoppableCheckWhenHelixOwnCheckFail() throws IOException {
     Map<String, Boolean> failedCheck = ImmutableMap.of("FailCheck", false);
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
-            HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient,
+            false, false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
               String instanceName, List<HealthCheck> healthChecks) {
             return failedCheck;
           }
         };
-
     String jsonContent = "";
     StoppableCheck actual =
         service.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
-
     Assert.assertEquals(actual.getFailedChecks().size(), failedCheck.size());
     Assert.assertFalse(actual.isStoppable());
     verifyZeroInteractions(_customRestClient);
@@ -104,8 +113,8 @@ public class TestInstanceService {
 
   @Test
   public void testGetInstanceStoppableCheckWhenCustomInstanceCheckFail() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -116,11 +125,9 @@ public class TestInstanceService {
     Map<String, Boolean> failedCheck = ImmutableMap.of("FailCheck", false);
     when(_customRestClient.getInstanceStoppableCheck(anyString(), anyMap()))
         .thenReturn(failedCheck);
-
     String jsonContent = "{\n" + "   \"param1\": \"value1\",\n" + "\"param2\": \"value2\"\n" + "}";
     StoppableCheck actual =
         service.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
-
     Assert.assertFalse(actual.isStoppable());
     Assert.assertEquals(actual.getFailedChecks().size(), failedCheck.size());
     verify(_customRestClient, times(1)).getInstanceStoppableCheck(any(), any());
@@ -129,8 +136,8 @@ public class TestInstanceService {
 
   @Test
   public void testGetInstanceStoppableCheckConnectionRefused() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -138,31 +145,25 @@ public class TestInstanceService {
             return Collections.emptyMap();
           }
         };
-
     when(_customRestClient.getInstanceStoppableCheck(anyString(), anyMap()))
         .thenThrow(IOException.class);
-
     Map<String, String> dataMap =
         ImmutableMap.of("instance", TEST_INSTANCE, "selection_base", "zone_based");
     String jsonContent = new ObjectMapper().writeValueAsString(dataMap);
     StoppableCheck actual =
         service.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
-
     int expectedFailedChecksSize = 1;
     String expectedFailedCheck =
         StoppableCheck.Category.CUSTOM_INSTANCE_CHECK.getPrefix() + TEST_INSTANCE;
-
     Assert.assertNotNull(actual);
     Assert.assertFalse(actual.isStoppable());
     Assert.assertEquals(actual.getFailedChecks().size(), expectedFailedChecksSize);
     Assert.assertEquals(actual.getFailedChecks().get(0), expectedFailedCheck);
   }
-
   @Test
   public void testCustomPartitionCheckWithSkipZKRead() throws IOException {
     // Let ZK result is health, but http request is unhealthy.
     // We expect the check fail if we skipZKRead.
-
     String testPartition = "PARTITION_0";
     String siblingInstance = "instance0.linkedin.com_1236";
     String jsonContent = "{\n" +
@@ -173,15 +174,12 @@ public class TestInstanceService {
     ZKHelixDataAccessor zkHelixDataAccessor =
         new ZKHelixDataAccessor(TEST_CLUSTER, mockAccessor);
     ZNRecord successPartitionReport = new ZNRecord(HelixDataAccessorWrapper.PARTITION_HEALTH_KEY);
-
     // Instance level check passed
     when(_customRestClient.getInstanceStoppableCheck(anyString(), anyMap())).thenReturn(
         Collections.singletonMap(TEST_INSTANCE, true));
-
     // Mocks for partition level
     when(mockAccessor.getChildNames(zkHelixDataAccessor.keyBuilder().liveInstances().getPath(), 2)).thenReturn(
         Arrays.asList(TEST_INSTANCE, siblingInstance));
-
     // Mock ZK record is healthy
     Map<String, String> partition0 = new HashMap<>();
     partition0.put(HelixDataAccessorWrapper.EXPIRY_KEY, String.valueOf(System.currentTimeMillis() + 100000L));
@@ -194,11 +192,9 @@ public class TestInstanceService {
         .healthReport(siblingInstance, HelixDataAccessorWrapper.PARTITION_HEALTH_KEY)
         .getPath()), Arrays.asList(new Stat(), new Stat()), 0, false)).thenReturn(
         Arrays.asList(successPartitionReport, successPartitionReport));
-
     // Mock client call result is check fail.
     when(_customRestClient.getPartitionStoppableCheck(anyString(), anyList(), anyMap())).thenReturn(
         Collections.singletonMap(testPartition, false));
-
     // Mock data for InstanceValidationUtil
     ExternalView externalView = new ExternalView("TestResource");
     externalView.setState(testPartition, TEST_INSTANCE, "MASTER");
@@ -210,21 +206,20 @@ public class TestInstanceService {
         AccessOption.PERSISTENT)).thenReturn(MasterSlaveSMD.build().getRecord());
 
     // Valid data only from ZK, pass the check
-    InstanceService instanceServiceReadZK =
-        new MockInstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, false,
-            false);
+    MockMaintenanceManagementService instanceServiceReadZK =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, false,
+            false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
     StoppableCheck stoppableCheck =
         instanceServiceReadZK.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
     Assert.assertTrue(stoppableCheck.isStoppable());
 
     // Even ZK data is valid. Skip ZK read should fail the test.
-    InstanceService instanceServiceWithoutReadZK =
-        new MockInstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, true,
-            false);
+    MockMaintenanceManagementService instanceServiceWithoutReadZK =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, true,
+            false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
     stoppableCheck = instanceServiceWithoutReadZK.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
     Assert.assertFalse(stoppableCheck.isStoppable());
   }
-
   /*
    * Tests stoppable check api when all checks query is enabled. After helix own check fails,
    * the subsequent checks should be performed.
@@ -235,13 +230,12 @@ public class TestInstanceService {
     BaseDataAccessor<ZNRecord> mockAccessor = mock(ZkBaseDataAccessor.class);
     ZKHelixDataAccessor zkHelixDataAccessor =
         new ZKHelixDataAccessor(TEST_CLUSTER, mockAccessor);
-
     when(mockAccessor.getChildNames(zkHelixDataAccessor.keyBuilder().liveInstances().getPath(), 2))
         .thenReturn(Arrays.asList(TEST_INSTANCE, siblingInstance));
 
     Map<String, Boolean> instanceHealthFailedCheck = ImmutableMap.of("FailCheck", false);
-    InstanceService service =
-        new InstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, true, true,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, true, true,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -249,31 +243,26 @@ public class TestInstanceService {
             return instanceHealthFailedCheck;
           }
         };
-
     when(_customRestClient.getInstanceStoppableCheck(anyString(), anyMap()))
         .thenReturn(ImmutableMap.of("FailCheck", false));
     when(_customRestClient.getPartitionStoppableCheck(anyString(), anyList(), anyMap()))
         .thenReturn(ImmutableMap.of("FailCheck", false));
-
     StoppableCheck actual = service.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, "");
     List<String> expectedFailedChecks =
         Arrays.asList("HELIX:FailCheck", "CUSTOM_INSTANCE_HEALTH_FAILURE:FailCheck");
-
     Assert.assertEquals(actual.getFailedChecks(), expectedFailedChecks);
     Assert.assertFalse(actual.isStoppable());
-
     // Verify the subsequent checks are called
     verify(_configAccessor, times(1)).getRESTConfig(anyString());
     verify(_customRestClient, times(1)).getInstanceStoppableCheck(anyString(), anyMap());
     verify(_customRestClient, times(2))
         .getPartitionStoppableCheck(anyString(), nullable(List.class), anyMap());
   }
-
   // TODO re-enable the test when partition health checks get decoupled
   @Test(enabled = false)
   public void testGetInstanceStoppableCheckWhenPartitionsCheckFail() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -281,29 +270,14 @@ public class TestInstanceService {
             return Collections.emptyMap();
           }
         };
-
     // partition is health on the test instance but unhealthy on the sibling instance
     when(_customRestClient.getInstanceStoppableCheck(anyString(), anyMap()))
         .thenReturn(Collections.emptyMap());
     String jsonContent = "{\n" + "   \"param1\": \"value1\",\n" + "\"param2\": \"value2\"\n" + "}";
     StoppableCheck actual =
         service.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
-
     Assert.assertFalse(actual.isStoppable());
     verify(_customRestClient, times(1)).getInstanceStoppableCheck(any(), any());
   }
 
-  private static class MockInstanceServiceImpl extends InstanceServiceImpl {
-    MockInstanceServiceImpl(ZKHelixDataAccessor dataAccessor, ConfigAccessor configAccessor,
-        CustomRestClient customRestClient, boolean skipZKRead, boolean continueOnFailures) {
-      super(dataAccessor, configAccessor, customRestClient, skipZKRead, continueOnFailures,
-          HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
-    }
-
-    @Override
-    protected Map<String, Boolean> getInstanceHealthStatus(String clusterId, String instanceName,
-        List<InstanceService.HealthCheck> healthChecks) {
-      return Collections.emptyMap();
-    }
-  }
 }
